Rentail. Ren. Tail. I was wondering why Text wouldn't appear. So I don't get really tight static analyse or compiler warnings that I can tell just yet. Will ahve to look into more code analysis option in WebStorm. I could rebase my awesome typo away, but I'm okay with the shame. Now ready to continue on with the helper portion of the tutorial.
Note of awesome: Ember Server will automatically detect changes in the file structure of the project that the ember-cli owns. It'll then recompile right then, so I'm running the server on a standalone powershell terminal running in administrator mode. Admin mode is for performance reasons, windows is pretty bad at dealing with these projects without some special configurations. The good part is that those configs are not especially difficult.