Okay, we've implemented the functions and properties. Of specific note are the autoComplete and search actions. These actions do not have their implementation set in the component, but rather are passed in as properties on the component under the filter property. The pattern is identified by the tutorial as closure actions. The details of how exactly this pattern is implemented is a bit unclear, but it's identified as a closure and has a structure that resembles inversion of control.
Video and some reading seems to suggest it allows us to decouple components from action dependencies. Previously, the component that required an action implementation would search up the heirarchy of components until it reached the controller which would house the implementation to run. With action closures, the controller can pass the required function into the component when it will be needed, and nested components no longer have to know about and implement helper chains to map to a function.